backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-multipart
pydantic[email]
python-jose
passlib[bcrypt]
pandas